requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pymongo==4.10.1
python-dotenv==1.0.1
transformers==4.45.1
python-multipart==0.0.12
PyPDF2==3.0.1
